# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation for Laravel API project",
        "description": "API documentation for Laravel API project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get all posts",
                "operationId": "ae6cf15376f3c11cd1245d604222ac15",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (by default = 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/posts?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/posts?page=1"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/posts"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": "http://localhost:8000/api/v1/posts?page=2"
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 6
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new post",
                "operationId": "b26b72502d82bdc3882283fc41fa26ef",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image",
                                    "category_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Title",
                                    "image": "image.png",
                                    "category_id": 1,
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post successful created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/{post}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post",
                "operationId": "da117e08b9b4071e22908ddbe383eff2",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image",
                                    "category_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Title",
                                    "image": "image.png",
                                    "category_id": 1,
                                    "user_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post successful updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "title": "Title",
                                    "image": "image.png",
                                    "category_id": 1,
                                    "user_id": 1,
                                    "created_at": "2024-01-01T00:00:00.000000Z",
                                    "updated_at": "2024-01-01T00:00:00.000000Z"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post",
                "operationId": "9f01a297017a4eee9c2f0f87da5f689f",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post successful deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/posts/employee/{employee}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get employee posts",
                "operationId": "8f0a50ab83fe5fd14f4801a8ef9e76e4",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/posts?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/posts?page=1"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/posts"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": "http://localhost:8000/api/v1/posts?page=2"
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 6
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Post": {
                "required": [
                    "id",
                    "title",
                    "image",
                    "category_id",
                    "user_id",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "title": "Title",
                    "image": "image.png",
                    "category_id": 1,
                    "user_id": 1,
                    "created_at": "2024-01-01T00:00:00.000000Z",
                    "updated_at": "2024-01-01T00:00:00.000000Z"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}